feat: M46 - Technical debt cleanup, full quality gates, and production baseline hardening
Backend:
- Fix SQLAlchemy model relationships (muted_by_users, followed_by_users)
- Add ArticleStatus.ANALYZED enum value
- Add Source.source_type field with SourceType enum
- Implement CompanyRepository and TopicRepository with get_by_slug
- Remove deprecated extract_article dependency
- Add type annotations to use cases
- Create missing evaluation module (confidence, evidence, contradiction)
- Fix migration conflicts (source_type in initial schema)
- Fix personalized feed tie-breaking stability
- Fix relevance_score type in feed response
- Remove unused type: ignore comments
- Add execute method to AnalyzeAndMaterializeUseCase stub

Frontend:
- Add missing notification types and API methods
- Fix NotificationBell, NotificationsPage, NotificationPreferencesPage types

Quality gates:
- Backend tests: 1471 passed
- Frontend tests: 25 passed
- Ruff check/format: all passed
- MyPy: success (no issues)
- TypeScript: pass
- Frontend lint: pass
- Coverage: 84.06%

+++ src/ai_news_digest/application/dto/user_preference/feed.py
@@ -24,7 +24,7 @@ class PersonalizedFeedItemResponse(BaseModel):
     article_count: int = 0
     source_count: int = 0
     relevance_reasons: list[str] = Field(default_factory=list)
-    relevance_score: int = 0
+    relevance_score: float = 0.0
     is_fallback: bool = False
     correction_signals: bool = False
     contradiction_signals: bool = False
